Use notifyPropertyChange on content instead of proxy
Changes done in Ember 3.13 around tracked properties broke this add-on. When a value of property changes on a buffer it somehow doesn't propagate the change.
In other words, if you have an <input> where you modify the value and you render the value in other places then it doesn't refresh.
Notifying a property change on content instead of proxy (this) seems to solve the issue.