Add acceptance tests
Using an acceptance test for testing the buffered proxy helps us in understanding if the addon plays well with the whole Ember framework. We see if views get re-render. We are also using Ember's internals instead of mocking them in unit tests. Thanks to that we should be able to tell if something changes when we deal with higher version of Ember.