🐞 Conditionally choose notifyPropertyChange context based on Ember version
Somehow, Ember v3.13.x introduced a change that caused
ember-buffered-proxy to stop working when calling notifyPropertyChange
to update references to a buffered object.

As a result of this change, a "fix" was introduced to
ember-buffered-proxy in
https://github.com/yapplabs/ember-buffered-proxy/commit/46d00d71cc85726025e1c385892d59078640914c
and https://github.com/yapplabs/ember-buffered-proxy/commit/3f63beb49185f7c377104689deb6395d93d1b8b4 which changed the way notifyPropertyChange
was called from:

```js
notifyPropertyChange(this, key);
```

to:

```js
notifyPropertyChange(content, key);
```

This seemed to fix the problem, although it doesn't conceptually make
sense as to why (which is slightly worrying).

Fast forward to Ember v3.20.x and something else has changed in Ember
source which has caused ember-buffered-proxy to stop working again. Or,
more likely, whatever was introduced in Ember 3.13.x has been patched so
that notifyPropertyChange, once again, requires `this` to be passed to
it, as was originally the case.

This commit adds some version checking to ensure that
notifyPropertyChange is called with the correct context depending on
what version of Ember is being used.